requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-appconf==1.0.4
django-compressor==2.4
django-sass-processor==0.8
gunicorn==20.0.4
idna==2.9
libsass==0.19.4
python-dotenv==0.13.0
pytz==2019.3
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
